+++ ReactUI/src/ui/components/materialUI/converter/FormVideoConverter.jsx
@@ -1,49 +1,10 @@
 import React from "react";
-import Paper from "@material-ui/core/Paper";
-import Grid from "@material-ui/core/Grid";
-import { makeStyles } from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
-import InputLabel from "@material-ui/core/InputLabel";
-import MenuItem from "@material-ui/core/MenuItem";
-import FormControl from "@material-ui/core/FormControl";
-import Select from "@material-ui/core/Select";
 import axios from "axios";
-
-import Card from "@material-ui/core/Card";
-import CardContent from "@material-ui/core/CardContent";
-import { CardHeader } from "@material-ui/core";
-
-import Table from "@material-ui/core/Table";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
-import TableContainer from "@material-ui/core/TableContainer";
-import TableHead from "@material-ui/core/TableHead";
-import TableRow from "@material-ui/core/TableRow";
-
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Checkbox from "@material-ui/core/Checkbox";
-
-const useStyles = makeStyles(() => ({
-	card: {
-		height: "100%",
-	},
-	input: {
-		display: "none",
-	},
-	root: {
-		width: 300,
-	},
-	title: {
-		color: "white",
-		fontSize: 12,
-		backgroundColor: "#3a4651",
-		//fontStyle: "italic",
-	},
-}));
+import VideoForm from "./VideoForm";
+import TableVideoForm from "./TableVideoForm";
 
 const FormVideoConveter = () => {
 	const urlML = "http://localhost:8080/imageFinder";
-	const classes = useStyles();
 
 	const [data, setResponse] = React.useState([]);
 	const [outputFormat, setOutputFormat] = React.useState("");
@@ -101,500 +62,33 @@ const FormVideoConveter = () => {
 	return (
 		<div>
 			<form onSubmit={submitFormVideo}>
-				<Grid container spacing={3} xs={12}>
-					<Grid item xs={12}>
-						<Card className={classes.card}>
-							<CardContent>
-								<input
-									accept="video/"
-									className={classes.input}
-									id="contained-button-file"
-									multiple
-									type="file"
-									required
-								/>
-								<label htmlFor="contained-button-file">
-									<Button
-										variant="contained"
-										color="primary"
-										component="span"
-									>
-										Upload
-									</Button>
-								</label>
-							</CardContent>
-						</Card>
-					</Grid>
-					<Grid item xs={7}>
-						<Card className={classes.card}>
-							<CardHeader
-								className={classes.title}
-								title="Video Format"
-								titleTypographyProps={{ variant: "h6" }}
-							/>
-							<CardContent>
-								<Grid container spacing={3}>
-									<Grid item xs>
-										<FormControl
-											variant="outlined"
-											className={classes.formControl}
-											fullWidth
-										>
-											<InputLabel id="demo-simple-select-outlined-label">
-												Output Format
-											</InputLabel>
-											<Select
-												labelId="demo-simple-select-outlined-label"
-												id="demo-simple-select-outlined"
-												value={outputFormat}
-												onChange={(e) =>
-													setOutputFormat(
-														e.target.value
-													)
-												}
-												label="Output Format"
-												required
-											>
-												<MenuItem value="">
-													<em>-</em>
-												</MenuItem>
-												<MenuItem value={".mp4"}>
-													.mp4
-												</MenuItem>
-												<MenuItem value={".mov"}>
-													.mov
-												</MenuItem>
-												<MenuItem value={".wmv"}>
-													.wmv
-												</MenuItem>
-												<MenuItem value={".avi"}>
-													.avi
-												</MenuItem>
-												<MenuItem value={".flv"}>
-													.flv
-												</MenuItem>
-												<MenuItem value={".mkv"}>
-													.mkv
-												</MenuItem>
-												<MenuItem value={".webm"}>
-													.webm
-												</MenuItem>
-											</Select>
-										</FormControl>
-									</Grid>
-									<Grid item xs>
-										<FormControl
-											variant="outlined"
-											className={classes.formControl}
-											fullWidth
-										>
-											<InputLabel id="demo-simple-select-outlined-label">
-												Fps
-											</InputLabel>
-											<Select
-												labelId="demo-simple-select-outlined-label"
-												id="demo-simple-select-outlined"
-												value={ratio}
-												onChange={(e) =>
-													setRatio(e.target.value)
-												}
-												label="Fps"
-												required
-											>
-												<MenuItem value="">
-													<em>-</em>
-												</MenuItem>
-												{[...Array(24)].map((x, i) => (
-													<MenuItem value={i + 1}>
-														{i + 1}
-													</MenuItem>
-												))}
-											</Select>
-										</FormControl>
-									</Grid>
-									<Grid item xs>
-										<FormControl
-											variant="outlined"
-											className={classes.formControl}
-											fullWidth
-										>
-											<InputLabel id="demo-simple-select-outlined-label">
-												Scale
-											</InputLabel>
-											<Select
-												labelId="demo-simple-select-outlined-label"
-												id="demo-simple-select-outlined"
-												value={scale}
-												onChange={(e) =>
-													setScale(e.target.value)
-												}
-												label="Scale"
-												required
-											>
-												<MenuItem value="">
-													<em>-</em>
-												</MenuItem>
-												<MenuItem value={"1280x720"}>
-													1280 x 720
-												</MenuItem>
-												<MenuItem value={"854x480"}>
-													854 x 480
-												</MenuItem>
-												<MenuItem value={"640x360"}>
-													640 x 360
-												</MenuItem>
-												<MenuItem value={"426x240"}>
-													426 x 240
-												</MenuItem>
-												<MenuItem value={"256x144"}>
-													256 x 144
-												</MenuItem>
-											</Select>
-										</FormControl>
-									</Grid>
-								</Grid>
-								<Grid container spacing={6}>
-									<Grid item xs>
-										<FormControl
-											variant="outlined"
-											className={classes.formControl}
-											fullWidth
-										>
-											<InputLabel id="demo-simple-select-outlined-label">
-												Audio format
-											</InputLabel>
-											<Select
-												labelId="demo-simple-select-outlined-label"
-												id="demo-simple-select-outlined"
-												value={audioFormat}
-												onChange={(e) =>
-													setAudioFormat(
-														e.target.value
-													)
-												}
-												label="Audio format"
-												required
-											>
-												<MenuItem value="">
-													<em>-</em>
-												</MenuItem>
-												<MenuItem value={".mp3"}>
-													.mp3
-												</MenuItem>
-												<MenuItem value={".mp2"}>
-													.mp2
-												</MenuItem>
-												<MenuItem value={".wma"}>
-													.wma
-												</MenuItem>
-											</Select>
-										</FormControl>
-									</Grid>
-									<Grid item xs>
-										<FormControl
-											variant="outlined"
-											className={classes.formControl}
-											fullWidth
-										>
-											<InputLabel id="demo-simple-select-outlined-label">
-												Quality
-											</InputLabel>
-											<Select
-												labelId="demo-simple-select-outlined-label"
-												id="demo-simple-select-outlined"
-												value={quality}
-												onChange={(e) =>
-													setQuality(e.target.value)
-												}
-												label="Quality"
-												required
-											>
-												{[...Array(31)].map((x, i) => (
-													<MenuItem value={i + 1}>
-														{i + 1}
-													</MenuItem>
-												))}
-											</Select>
-										</FormControl>
-									</Grid>
-								</Grid>
-							</CardContent>
-						</Card>
-					</Grid>
-					<Grid item xs={5}>
-						<Card className={classes.card}>
-							<CardHeader
-								className={classes.title}
-								title="Flip and rotate"
-								titleTypographyProps={{ variant: "h6" }}
-							/>
-							<CardContent>
-								<Grid container spacing={3}>
-									<Grid item xs>
-										<FormControl
-											variant="outlined"
-											className={classes.formControl}
-											fullWidth
-										>
-											<InputLabel id="demo-simple-select-outlined-label">
-												Rotate
-											</InputLabel>
-											<Select
-												labelId="demo-simple-select-outlined-label"
-												id="demo-simple-select-outlined"
-												value={angle}
-												onChange={(e) =>
-													setAngle(e.target.value)
-												}
-												label="Scale"
-												required
-											>
-												<MenuItem value={"90"}>
-													90
-												</MenuItem>
-												<MenuItem value={"180"}>
-													180
-												</MenuItem>
-												<MenuItem value={"270"}>
-													270
-												</MenuItem>
-											</Select>
-										</FormControl>
-									</Grid>
-								</Grid>
-								<Grid container spacing={6}>
-									<Grid item xs>
-										<FormControlLabel
-											control={
-												<Checkbox
-													checked={hflip}
-													onChange={(e) => {
-														sethFlip(
-															e.target.checked
-														);
-													}}
-													name="flipHorizon"
-													color="primary"
-												/>
-											}
-											label="Flip Horizon"
-										/>
-									</Grid>
-
-									<Grid item xs>
-										<FormControlLabel
-											control={
-												<Checkbox
-													checked={vflip}
-													onChange={(e) => {
-														setvFlip(
-															e.target.checked
-														);
-													}}
-													name="flipVertical"
-													color="primary"
-												/>
-											}
-											label="Flip Vertical"
-										/>
-									</Grid>
-								</Grid>
-							</CardContent>
-						</Card>
-					</Grid>
-				</Grid>
-
-				<Grid container spacing={3}>
-					<Grid item xs={6}>
-						<Card className={classes.card}>
-							<CardHeader
-								className={classes.title}
-								title="Frames"
-								titleTypographyProps={{ variant: "h6" }}
-							/>
-							<CardContent>
-								<Grid container spacing={6}>
-									<Grid item xs>
-										<FormControl
-											variant="outlined"
-											className={classes.formControl}
-											fullWidth
-										>
-											<InputLabel id="demo-simple-select-outlined-label">
-												Frames
-											</InputLabel>
-											<Select
-												labelId="demo-simple-select-outlined-label"
-												id="demo-simple-select-outlined"
-												label="Frames"
-												value={frameScale}
-												onChange={(e) =>
-													setFrameScale(
-														e.target.value
-													)
-												}
-												required
-											>
-												<MenuItem value={"854"}>
-													854
-												</MenuItem>
-												<MenuItem value={"640"}>
-													640
-												</MenuItem>
-												<MenuItem value={"520"}>
-													520
-												</MenuItem>
-												<MenuItem value={"426"}>
-													426
-												</MenuItem>
-												<MenuItem value={"360"}>
-													360
-												</MenuItem>
-												<MenuItem value={"256"}>
-													256
-												</MenuItem>
-											</Select>
-										</FormControl>
-									</Grid>
-									<Grid item xs>
-										<FormControlLabel
-											control={
-												<Checkbox
-													checked={obtainFrames}
-													onChange={(e) => {
-														setObtainFrames(
-															e.target.checked
-														);
-													}}
-													name="obtainFrames"
-													color="primary"
-												/>
-											}
-											label="Obtain frames"
-										/>
-									</Grid>
-								</Grid>
-							</CardContent>
-						</Card>
-					</Grid>
-					<Grid item xs={6}>
-						<Card className={classes.card}>
-							<CardHeader
-								className={classes.title}
-								title="Obtain Audio"
-								titleTypographyProps={{ variant: "h6" }}
-							/>
-							<CardContent>
-								<Grid container spacing={6}>
-									<Grid item xs>
-										<FormControl
-											variant="outlined"
-											className={classes.formControl}
-											fullWidth
-										>
-											<InputLabel id="demo-simple-select-outlined-label">
-												Audio format
-											</InputLabel>
-											<Select
-												labelId="demo-simple-select-outlined-label"
-												id="demo-simple-select-outlined"
-												value={extractAudioFormat}
-												onChange={(e) =>
-													setExtractAudioFormat(
-														e.target.value
-													)
-												}
-												label="Audio format"
-												required
-											>
-												<MenuItem value={".mp3"}>
-													.mp3
-												</MenuItem>
-												<MenuItem value={".mp2"}>
-													.mp2
-												</MenuItem>
-												<MenuItem value={".wma"}>
-													.wma
-												</MenuItem>
-											</Select>
-										</FormControl>
-									</Grid>
-									<Grid item xs>
-										<FormControlLabel
-											control={
-												<Checkbox
-													checked={obtainAudio}
-													onChange={(e) => {
-														setObtainAudio(
-															e.target.checked
-														);
-													}}
-													name="obtainAudio"
-													color="primary"
-												/>
-											}
-											label="Obtain Audio"
-										/>
-									</Grid>
-								</Grid>
-							</CardContent>
-						</Card>
-					</Grid>
-					<Grid item xs={12}>
-						<Card className={classes.card}>
-							<CardHeader
-								className={classes.title}
-								title="Results"
-								titleTypographyProps={{ variant: "h6" }}
-							/>
-							<CardContent>
-								<Grid container spacing={3}>
-									<Grid item xs>
-										<Button
-											type="submit"
-											variant="contained"
-											color="primary"
-										>
-											Convert
-										</Button>
-									</Grid>
-								</Grid>
-								<Grid container spacing={6}>
-									<Grid item xs>
-										<TableContainer component={Paper}>
-											<Table
-												className={classes.table}
-												size="small"
-												aria-label="a dense table"
-											>
-												<TableHead>
-													<TableRow>
-														<TableCell align="center">
-															Name
-														</TableCell>
-														<TableCell align="center">
-															Link
-														</TableCell>
-													</TableRow>
-												</TableHead>
-												<TableBody>
-													<TableRow>
-														<TableCell align="center">
-															Name
-														</TableCell>
-														<TableCell align="center">
-															Link
-														</TableCell>
-													</TableRow>
-												</TableBody>
-											</Table>
-										</TableContainer>
-									</Grid>
-								</Grid>
-							</CardContent>
-						</Card>
-					</Grid>
-				</Grid>
+				<VideoForm
+					ratio={ratio}
+					scale={scale}
+					quality={quality}
+					angle={angle}
+					vflip={vflip}
+					hflip={hflip}
+					outputFormat={outputFormat}
+					audioFormat={audioFormat}
+					obtainFrames={obtainFrames}
+					frameScale={frameScale}
+					obtainAudio={obtainAudio}
+					extractAudioFormat={extractAudioFormat}
+					setOutputFormat={setOutputFormat}
+					setRatio={setRatio}
+					setScale={setScale}
+					setAudioFormat={setAudioFormat}
+					setQuality={setQuality}
+					setAngle={setAngle}
+					sethFlip={sethFlip}
+					setvFlip={setvFlip}
+					setFrameScale={setFrameScale}
+					setObtainFrames={setObtainFrames}
+					setExtractAudioFormat={setExtractAudioFormat}
+					setObtainAudio={setObtainAudio}
+				/>
+				<TableVideoForm />
 			</form>
 		</div>
 	);
